Refactor typeChecker
